Removed old create_file
