+++ backend/services/create_file.py
@@ -1,79 +1,64 @@
 import asyncio
-import csv
 import json
 import os
 import sys
 
-from api.routers.nba import (
-    get_all_players,
-    get_player_career_stats,
-    get_player_game_log,
-    get_player_info,
-)
+import httpx
+import pandas as pd
 
-# TODO: Rewrite Create File
-# Look to use the api endpoints to fetch the 
-# data rather then trying to call the function
+from api.routers.nba import get_all_players
+
+BASE_URL = "http://127.0.0.1:8000"
+
 
 def create_file(data, file_name, output_format):
-    data_directory = "../data"
-    file_path = os.path.join(data_directory, file_name)
+    data_dir = "../data"
+    file_path = os.path.join(data_dir, file_name)
 
-    os.makedirs(data_directory, exist_ok=True)
+    os.makedirs(data_dir, exist_ok=True)
 
     if output_format == "json":
         with open(file_path, "w") as file:
-            json.dump(json.loads(data), file, indent=4)
+            json.dump(data, file, indent=4)
     elif output_format == "csv":
-        if isinstance(data, str):
-            data = json.loads(data)
-        keys = data[0].keys() if data else []
-        with open(file_path, "w", newline="") as output_file:
-            dict_writer = csv.DictWriter(output_file, fieldnames=keys)
-            dict_writer.writeheader()
-            dict_writer.writerows(data)
+        with open(file_path, "w") as file:
+            df = pd.DataFrame(data)
+            df.to_csv(file, index=False)
 
 
-# NOTE: Mapping of command-line arguments to functions in nba_service
-# Add new functions here when created so that we can create json files for those functions in nba_service.py
-function_mapping = {
-    "all_players": get_all_players,
-    "player_career_stats": get_player_career_stats,
-    "player_game_log": get_player_game_log,
-    "player_info": get_player_info,
-}
+async def fetch_data(endpoint, params):
+    async with httpx.AsyncClient() as client:
+        response = await client.get(endpoint, params=params)
+        response.raise_for_status()
+        return response.json()
 
 
 async def main():
-    if len(sys.argv) < 3:
-        print("Usage: python create_file.py <function> <file_name> [arguments]")
+    if len(sys.argv) < 4:
+        print("Usage: python create_file.py <sports_league> <type_of_data> <output_format> <args...>")
         sys.exit(1)
 
-    function_name = sys.argv[1]
-    output_format = sys.argv[2]
-    args = sys.argv[3:]
+    sports_league = sys.argv[1]
+    type_of_data = sys.argv[2]
+    output_format = sys.argv[3]
+    args = sys.argv[4:]
 
-    if function_name.startswith("get_"):
-        function_name = function_name.replace("get_", "")
+    endpoint = f"{BASE_URL}/{sports_league}/{type_of_data}/{'/'.join(args)}"
+    params = {"output_format": output_format}
 
-    if function_name not in function_mapping:
-        print(
-            f"Invalid function name. Valid options are: {list(function_mapping.keys())}"
-        )
-        sys.exit(1)
+    try:
+        data = await fetch_data(endpoint, params)
+        file_name = f"{type_of_data}_{'_'.join(args)}.{output_format}"
 
-    if args:
-        data = await function_mapping[function_name](*args)
-        file_name = f"{function_name}_{'_'.join(args)}.{output_format}"
-    else:
-        data = await function_mapping[function_name](output_format)
-        file_name = f"{function_name}.{output_format}"
-
-    if data is not None:
-        create_file(data, file_name, output_format)
-        print(f"File created: {file_name}")
-    else:
-        print(f"No data returned for {function_name} with arguments{args}.")
+        if data is not None:
+            create_file(data, file_name, output_format)
+            print(f"File created: {file_name}")
+        else:
+            print(f"No data returned for {type_of_data} with arguments {args}.")
+    except httpx.HTTPStatusError as e:
+        print(f"HTTP Exception: {e.response.status_code} - {e.response.text}")
+    except Exception as e:
+        print(f"Unexpected error: {str(e)}")
 
 
 if __name__ == "__main__":
